Separate content_providers.py + fixed some imports

+++ src/content_providers/fixed_profiles_provider.py
@@ -1,15 +1,13 @@
 import torch
 import pandas as pd
 
+from neural_collaborative_filtering.content_providers import ContentProvider
 from globals import item_metadata_file, user_embeddings_file
-
-from neural_collaborative_filtering.datasets.base import ContentProvider
 
 
 class FixedProfilesProvider(ContentProvider):
     # noinspection PyTypeChecker
     def __init__(self):
-        # TODO: use global to load them here or take as args?
         self.metadata: pd.DataFrame = pd.read_hdf(item_metadata_file + '.h5')
         self.user_embeddings: pd.DataFrame = pd.read_hdf(user_embeddings_file + '.h5')
 
@@ -23,7 +21,7 @@ class FixedProfilesProvider(ContentProvider):
         # Note: Old way for non fixed embeddings.
         # This is for just-in-time user embedding creation. Better to do pre-do this for the whole class.
         avg_rating = user_ratings['rating'].mean()
-        return torch.FloatTensor(((user_ratings['rating'] - avg_rating) * self.metadata.loc[user_ratings['movieId']].values).mean(axis=0))   # TODO: might be wrong
+        return ((user_ratings['rating'] - avg_rating) * self.metadata.loc[user_ratings['movieId']].values).mean(axis=0)   # TODO: might be wrong
 
     def get_num_items(self):                     # TODO: what if we store more than are in our samples?
         return self.metadata.shape[0]
